requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements.in
#
altair==4.2.0
    # via streamlit
attrs==22.1.0
    # via jsonschema
backports-zoneinfo==0.2.1
    # via
    #   pytz-deprecation-shim
    #   tzlocal
blinker==1.5
    # via streamlit
cachetools==5.2.0
    # via streamlit
certifi==2022.9.24
    # via requests
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via streamlit
commonmark==0.9.1
    # via rich
contourpy==1.0.6
    # via matplotlib
cycler==0.11.0
    # via matplotlib
decorator==5.1.1
    # via validators
entrypoints==0.4
    # via altair
fonttools==4.38.0
    # via matplotlib
gitdb==4.0.10
    # via gitpython
gitpython==3.1.29
    # via streamlit
htbuilder==0.6.1
    # via -r requirements.in
idna==3.4
    # via requests
importlib-metadata==5.1.0
    # via streamlit
importlib-resources==5.10.0
    # via jsonschema
jinja2==3.1.2
    # via
    #   altair
    #   pydeck
joblib==1.2.0
    # via
    #   scikit-learn
    #   scikit-surprise
jsonschema==4.17.3
    # via altair
kiwisolver==1.4.4
    # via matplotlib
markupsafe==2.1.1
    # via jinja2
matplotlib==3.6.2
    # via
    #   -r requirements.in
    #   wordcloud
more-itertools==9.0.0
    # via htbuilder
numpy==1.23.5
    # via
    #   altair
    #   contourpy
    #   matplotlib
    #   pandas
    #   pyarrow
    #   pydeck
    #   scikit-learn
    #   scikit-surprise
    #   scipy
    #   streamlit
    #   wordcloud
packaging==21.3
    # via
    #   matplotlib
    #   streamlit
pandas==1.5.2
    # via
    #   -r requirements.in
    #   altair
    #   streamlit
pillow==9.3.0
    # via
    #   -r requirements.in
    #   matplotlib
    #   streamlit
    #   wordcloud
pkgutil-resolve-name==1.3.10
    # via jsonschema
protobuf==3.20.3
    # via streamlit
pyarrow==10.0.1
    # via streamlit
pydeck==0.8.0
    # via streamlit
pygments==2.13.0
    # via rich
pympler==1.0.1
    # via streamlit
pyparsing==3.0.9
    # via
    #   matplotlib
    #   packaging
pyrsistent==0.19.2
    # via jsonschema
python-dateutil==2.8.2
    # via
    #   matplotlib
    #   pandas
    #   streamlit
pytz==2022.6
    # via pandas
pytz-deprecation-shim==0.1.0.post0
    # via tzlocal
requests==2.28.1
    # via streamlit
rich==12.6.0
    # via streamlit
scikit-learn==1.1.3
    # via -r requirements.in
scikit-surprise==1.1.3
    # via surprise
scipy==1.9.3
    # via
    #   -r requirements.in
    #   scikit-learn
    #   scikit-surprise
semver==2.13.0
    # via streamlit
six==1.16.0
    # via python-dateutil
smmap==5.0.0
    # via gitdb
streamlit==1.15.1
    # via -r requirements.in
surprise==0.1
    # via -r requirements.in
threadpoolctl==3.1.0
    # via scikit-learn
toml==0.10.2
    # via streamlit
toolz==0.12.0
    # via altair
tornado==6.2
    # via streamlit
tqdm==4.64.1
    # via -r requirements.in
typing-extensions==4.4.0
    # via
    #   rich
    #   streamlit
tzdata==2022.6
    # via pytz-deprecation-shim
tzlocal==4.2
    # via streamlit
urllib3==1.26.13
    # via requests
validators==0.20.0
    # via streamlit
watchdog==2.1.9
    # via streamlit
wordcloud==1.8.2.2
    # via -r requirements.in
zipp==3.11.0
    # via
    #   importlib-metadata
    #   importlib-resources
